+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.7] 2021-03-16
+
+### Added
+
+-   `onceAsPromise()` method that returns a Promise that resolves the next time
+    the specified event is emitted.
+
 ## [0.0.6] 2021-03-09
 
-## Added
+### Added
 
 -   Support for `unique symbol` event names to the `EventSource.subscribe()` method.
 
@@ -76,6 +83,7 @@ TODO:
 -   Write good README content (sorry; source code documentation will have to suffice for now).
 -   Use rollup to bundle code?
 
+[0.0.7]: https://github.com/UselessPickles/strictly-typed-events/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/UselessPickles/strictly-typed-events/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/UselessPickles/strictly-typed-events/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/UselessPickles/strictly-typed-events/compare/v0.0.3...v0.0.4